backend/pyproject.toml
[tool.poetry]
name = "reflect"
version = "0.1.0"
description = "AI agent that analyzes raw video/audio footage and edits it to a music beat"
authors = ["Edwin Luo"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
librosa = "^0.11.0"
python-dotenv = "^1.0"
pydantic = "^2.0"
openai-agents = {extras = ["litellm"], version = "^0.6.5"}
pymongo = "^4.16.0"
opentimelineio = "^0.18.1"
opencv-python = "^4.12.0.88"
faster-whisper = "^1.2.1"
pydantic-mongo = "^3.1.0"
motor = "^3.7.1"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.20"
websockets = "^14.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14"
pyright = "^1.1"
pytest = "^9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# =============================================================================
# RUFF CONFIGURATION (Strict Settings)
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "scratch"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # ruff-specific rules
    "PERF",   # perflint
    "FBT",    # flake8-boolean-trap
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "DJ",     # flake8-django
    "EM",     # flake8-errmsg
    "EXE",    # flake8-executable
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PYI",    # flake8-pyi
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "TID",    # flake8-tidy-imports
    "INT",    # flake8-gettext
    "FLY",    # flynt
    "LOG",    # flake8-logging
    "N",      # pep8-naming
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "BLE",    # flake8-blind-except
    "ANN",    # flake8-annotations
    "D",      # pydocstyle
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring (conflicts with D211)
    "D213",   # Multi-line docstring summary should start at the second line (conflicts with D212)
    "COM812", # Trailing comma missing (conflicts with formatter)
    "ISC001", # Single line implicit string concatenation (conflicts with formatter)
    "T201",   # Allow print statements in scratch files
]

[tool.ruff.lint.per-file-ignores]
"src/*" = [
    "INP001", # Allow implicit namespace package (no __init__.py per style guide)
]
"src/core/*" = [
    "S603",   # Allow subprocess calls (ffmpeg execution)
    "S607",   # Allow partial executable paths (ffmpeg in PATH)
]
"src/agents/*" = [
    "S603",   # Allow subprocess calls (ffmpeg for audio extraction)
    "S607",   # Allow partial executable paths (ffmpeg in PATH)
]
"src/asset_annotator/*" = [
    "S603",   # Allow subprocess calls (ffprobe execution)
    "S607",   # Allow partial executable paths (ffprobe in PATH)
]
"scratch/*" = [
    "T201",   # Allow print statements
    "D103",   # Missing docstring in public function
    "S101",   # Allow assert statements
    "INP001", # Allow implicit namespace package (scratch folder)
    "ANN",    # Relax annotations for scratch files
    "PLR0913", # Too many arguments
    "PLR2004", # Magic values in comparisons
    "PLC0415", # Import not at top-level
]
"tests/*" = [
    "S101",   # Allow assert statements
    "D103",   # Missing docstring in public function
    "ANN",    # Relax annotations for tests
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# =============================================================================
# PYRIGHT CONFIGURATION
# =============================================================================
[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "All"
typeCheckingMode = "strict"
include = ["src", "scratch"]
exclude = [
    "**/__pycache__",
    ".venv",
    "venv",
]

# Core type checking (keep strict)
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
reportPrivateUsage = true
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
reportOverlappingOverload = true
reportUninitializedInstanceVariable = true
reportCallInDefaultInitializer = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportAssertAlwaysTrue = true
reportSelfClsParameterName = true
reportImplicitStringConcatenation = false
reportUnusedCallResult = false
reportPropertyTypeMismatch = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true
reportImplicitOverride = false

# Relaxed for untyped libraries (librosa, ffmpeg-python lack stubs)
# These generate noise when using libraries without type annotations
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportMissingTypeArgument = false
reportUnknownLambdaType = false

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
